refactor(ts-migration): move runtime leaves to typescript
